Fix some warnings generated by gcc 4.8 and clang

+++ Core/MIPS/ARM/ArmCompFPU.cpp
@@ -56,10 +56,10 @@ void Jit::Comp_FPU3op(u32 op)
 	case 2: { //F(fd) = F(fs) * F(ft); //mul
 		u32 nextOp = Memory::Read_Instruction(js.compilerPC + 4);
 		// Optimise possible if destination is the same
-		if (fd == ((nextOp>>6) & 0x1F)) {
+		if (fd == (int)((nextOp>>6) & 0x1F)) {
 			// VMUL + VNEG -> VNMUL
 			if (!strcmp(MIPSGetName(nextOp), "neg.s")) {
-				if (fd == ((nextOp>>11) & 0x1F)) {
+				if (fd == (int)((nextOp>>11) & 0x1F)) {
 					VNMUL(fpr.R(fd), fpr.R(fs), fpr.R(ft));
 					EatInstruction(nextOp);
 				}
